+++ bootique-metrics-healthchecks/src/main/java/io/bootique/metrics/health/HealthCheckOutcome.java
@@ -74,7 +74,7 @@ public class HealthCheckOutcome implements Comparable<HealthCheckOutcome> {
 
     /**
      * @return a {@link HealthCheckOutcome} corresponding to a "warning" state of the app, i.e. approaching critical.
-     * @since 0.26
+     * @since 1.0.RC1
      */
     public static HealthCheckOutcome warning(Throwable th) {
         return new HealthCheckOutcome(HealthCheckStatus.WARNING, null, th);
@@ -139,7 +139,7 @@ public class HealthCheckOutcome implements Comparable<HealthCheckOutcome> {
     /**
      * @return a new {@link HealthCheckOutcome} with all the information from this outcome plus extra metrics data
      * that was used to generate this outcome.
-     * @since 0.26
+     * @since 1.0.RC1
      */
     public HealthCheckOutcome withData(HealthCheckData<?> data) {
         HealthCheckOutcome outcomeWithData = new HealthCheckOutcome(status, message, error);
@@ -152,7 +152,7 @@ public class HealthCheckOutcome implements Comparable<HealthCheckOutcome> {
      * state. Kind of a metrics for health check.
      *
      * @return an optional extra data for this health check.
-     * @since 0.26
+     * @since 1.0.RC1
      */
     public Optional<HealthCheckData<?>> getData() {
         return data;