+++ bootique-metrics-healthchecks/src/main/java/io/bootique/metrics/health/check/DoubleRangeFactory.java
@@ -23,7 +23,7 @@ import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 @BQConfig
 public class DoubleRangeFactory {