+++ bootique-metrics-healthchecks/src/main/java/io/bootique/metrics/health/check/DurationRangeFactory.java
@@ -24,7 +24,7 @@ import io.bootique.annotation.BQConfigProperty;
 import io.bootique.value.Duration;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 @BQConfig
 public class DurationRangeFactory {