+++ bootique-metrics-healthchecks/src/main/java/io/bootique/metrics/health/check/PercentRangeFactory.java
@@ -23,7 +23,7 @@ import io.bootique.annotation.BQConfigProperty;
 import io.bootique.value.Percent;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class PercentRangeFactory {