+++ bootique-metrics/src/main/java/io/bootique/metrics/MetricNaming.java
@@ -27,7 +27,7 @@ import java.util.StringJoiner;
 /**
  * A helper class to name metrics across Bootique modules. Helps somewhat to enforce consistent naming convention. Note that
  *
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class MetricNaming {