+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #27 Support for unit of measurement in HealthCheckData
 * #28 Cleaning up APIs deprecated since <= 0.25
 * #29 DeferredHealthCheck - a health checking proxy for services not available on startup
+* #30 Metrics naming conventions
 
 ## 0.25